% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knappsack.R
\name{knapsackgit}
\alias{knapsackgit}
\title{Knapsack problem solver}
\usage{
knapsackgit(w, v, c)
}
\arguments{
\item{w:weights}{vector}

\item{v;value}{vector}

\item{c:capacity}{(numeric)}
}
\value{
Return the maximum of the w*v which are less than or equal to capacity
}
\description{
take the weights, values and maximum capacity as parameters
}
